Finalized nn_trainer structure and have started working on the genre guesser
