+++ .gitignore
@@ -1,0 +1,2 @@
+/Github-ignore/
+/venv/
