+++ NN-Genre-Guesser/genre-guesser.py
@@ -1,0 +1,17 @@
+"""
+This file contains the primary functions that allow the NN to guess what genres the inputted
+song might be related to.
+"""
+
+import keras
+
+
+def genre_guesser(h5_song_file):
+    """
+    The genre_guesser guesses what genres are related to the given song data.
+
+    :param h5_song_file:
+    :return: A dictionary with the percentile relationship to each genre in the NN
+    """
+    pass
+
